tests/wpt/web-platform-tests/portals/portals-activate-inside-portal.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <script>
    promise_test(async () => {
      var portal = document.createElement("portal");
      portal.src = "resources/portal-activate-inside-portal.html";
      let waitForMessage = new Promise((resolve, reject) => {
        portal.onmessage = e => resolve(e.data);
        document.body.appendChild(portal);
      });
      var error = await waitForMessage;
      assert_equals(error, "InvalidStateError");
    }, "activating a nested portal should throw an error");
  </script>
</body>
